Rename _assets to _includes to allow including arbitrary files
This requirement came about when trying to publish to GH pages, which
requires a CNAME file in the root of the repo when using custom domains.

Therefore better to make a generic system where we move all included
files out of the way of the docs, and dump them to the root of the site
on build.